tests/wpt/tests/css/css-masonry/tentative/parsing/masonry-track-computed.html

<!DOCTYPE html>
<title>CSS Masonry: masonry-track-* getComputedStyle()</title>
<link rel="author" title="Sam Davis Omekara Jr." href="mailto:samomekarajr@microsoft.com">
<link rel="help" href="https://tabatkins.github.io/specs/css-masonry/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<script src="/css/support/inheritance-testcommon.js"></script>
<div id="container">
  <div id="target"></div>
</div>
<script>
  // auto
  test_computed_value("masonry-track-end", "auto");
  test_computed_value("masonry-track-start", "auto");

  // <custom-ident>
  test_computed_value("masonry-track-end", "_9");
  test_computed_value("masonry-track-start", "-_π");

  // <integer> && <custom-ident>?
  test_computed_value("masonry-track-end", "-44 Z");
  test_computed_value("masonry-track-start", "+90 X", "90 X");

  // span && [ <integer> || <custom-ident> ]
  test_computed_value("masonry-track-end", "span 1 i", "span i");
  test_computed_value("masonry-track-start", "span 2");

  // Computed Value for `masonry-track` should be the shortest
  // serialization possible.
  test_computed_value("masonry-track", "auto / auto", "auto");
  test_computed_value("masonry-track", "auto", "auto");
  test_computed_value("masonry-track", "10 / auto", "10");
  test_computed_value("masonry-track", "10", "10");
  test_computed_value("masonry-track", "-10 / auto", "-10");
  test_computed_value("masonry-track", "-10", "-10");
  test_computed_value("masonry-track", "first / first", "first");
  test_computed_value("masonry-track", "first", "first");
  test_computed_value("masonry-track", "span 2 / auto", "span 2");
  test_computed_value("masonry-track", "span 2", "span 2");
  test_computed_value("masonry-track", "2 first / auto", "2 first");
  test_computed_value("masonry-track", "2 first", "2 first");
  test_computed_value("masonry-track", "span first / auto", "span first");
  test_computed_value("masonry-track", "span first", "span first");
  test_computed_value("masonry-track", "span 2 first / auto", "span 2 first");
  test_computed_value("masonry-track", "span 2 first", "span 2 first");
</script>
